Check stale read before saving vastaanottos
This fixes the problem that saving vastaanottos to VTS writes a log
entry to the relevant Valintatulos which causes a StaleReadException
when saving the rest of the Valintatulos data to sijoittelu-service.
Checking for stale reads before VTS call leaves a small window of
possibility for overwriting concurrent changes that happened during the
VTS call.